// src/domain/enums/PracticalSkill.ts
export enum PracticalSkill {
  Electrical = "Electrical",
  Plumbing = "Plumbing",
  Construction = "Construction",
  Painting = "Painting",
  Carpentry = "Carpentry",
  Gardening = "Gardening",
  Mechanics = "Mechanics",
  IT = "IT",
  GraphicDesign = "GraphicDesign",
  VideoEditing = "VideoEditing",
  Sewing = "Sewing",
  Cooking = "Cooking",
  Driving = "Driving",
  Health = "Health",
  Law = "Law",
  AccountingFinance = "AccountingFinance",
  Education = "Education",
  TranslationLanguages = "TranslationLanguages",
  Other = "Other",
}